fix(elements-core): make useResizer private. fix types (#2268)
