--- packages/elements-core/src/hooks/useResizer.ts
@@ -1,35 +1,0 @@
-import React, { useEffect, useRef, useState } from 'react';
-
-export function useResizer(sidebarWidth: number): [React.Ref<HTMLDivElement>, number, (event: any) => void] {
-  const sidebarRef = useRef<HTMLDivElement | null>(null);
-  const [isResizing, setIsResizing] = useState(false);
-  const [currentSidebarWidth, setCurrentSidebarWidth] = useState(sidebarWidth);
-
-  const startResizing = React.useCallback(() => {
-    setIsResizing(true);
-  }, []);
-
-  const stopResizing = React.useCallback(() => {
-    setIsResizing(false);
-  }, []);
-
-  const resize = React.useCallback(
-    mouseMoveEvent => {
-      if (isResizing) {
-        setCurrentSidebarWidth(mouseMoveEvent.clientX - sidebarRef.current!.getBoundingClientRect().left);
-      }
-    },
-    [isResizing],
-  );
-
-  useEffect(() => {
-    window.addEventListener('mousemove', resize);
-    window.addEventListener('mouseup', stopResizing);
-    return () => {
-      window.removeEventListener('mousemove', resize);
-      window.removeEventListener('mouseup', stopResizing);
-    };
-  }, [resize, stopResizing]);
-
-  return [sidebarRef, currentSidebarWidth, startResizing];
-}
